requirements.txt
aioredis==1.2.0
amqp==2.4.0
anyjson==0.3.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
celery-with-redis==3.0
certifi==2018.11.29
cffi==1.11.5
channels==2.1.6
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
daphne==2.2.4
Django==2.1
django-celery==3.2.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-redis==4.10.0
django-rest-swagger==2.2.0
django-timezone-field==3.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
flower==0.9.2
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
MarkupSafe==1.1.0
msgpack==0.6.1
numpy==1.15.4
openapi-codec==1.3.2
opencv-python==3.4.4.19
Pillow==5.4.1
pyasn1==0.4.5
pycparser==2.19
pycryptodomex==3.7.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.9.1
python-crontab==2.3.6
python-dateutil==2.7.5
pytz==2018.9
redis==3.0.1
requests==2.21.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
tornado==5.1.1
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
zope.interface==4.6.0
